% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logo.R
\name{logo}
\alias{logo}
\title{Insert logo from the package files}
\usage{
logo(x = "mnb_intezet")
}
\arguments{
\item{x}{file name of the logo (without `.png`)}
}
\value{
Exact location of the logo.
}
\description{
Insert logo from the package files
}
